Don't expect leal %edi in fast-isel-x32.ll
X32 psABI says "When a value of pointer type is returned or passed in a
register, bits 32 to 63 shall be zero".  We need to zero-extend the
lower 32 bits of a 64-bit register to 64 bits when passing a pointer.
fast-isel doesn't optimize out zero-extend.  Also x32 should use
LEA64_32r instead of LEA32r to avoid 0x67 address-size prefix.